pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nuwinter"
version = "1.0.0"
description = ""
authors = [
    {name = "Robert Stein", email = "rdstein@caltech.edu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "winterapi>=1.5.0",
    "jupyter",
    "matplotlib",
    "requests",
    "wintertoo>=1.5.0",
    "astropy",
    "matplotlib",
    "numpy",
    "nuztf>=3.0.0",
    "pandas",
    "pydantic>2.0.0",
    "python-dotenv",
    "photutils"
]

[project.urls]
Homepage = "https://github.com/robertdstein/nuwinter"

[tool.setuptools]
packages = ["nuwinter"]

[tool.coverage.run]
source = ["nuwinter"]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",
    # Don't complain about missing debug-only code:
    "def __repr__",
    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",
    "raise KeyError",
    "except KeyError",
    "raise ValueError",
    "except JSONDecodeError:",
    "raise requests.exceptions.RequestException",
    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if False:",
    'if __name__ == "__main__":',
    "err =",
    "logger.error",
    "raise"
    ]
ignore_errors = true

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = "88"
disable=["logging-fstring-interpolation"]
good-names=["ax", "ra", "df", "pi", "i"]
exclude-too-few-public-methods=["pydantic.*"]
extension-pkg-whitelist=["pydantic"]
